threadx_dashboard/requirements.txt
# Framework principal
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0

# Manipulation de données
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0

# Backend et serveur
flask==3.0.0
flask-cors==4.0.0

# Sécurité et authentification
bcrypt==4.1.1
python-dotenv==1.0.0
PyJWT==2.8.1

# Base de données (optionnel)
sqlalchemy==2.0.23

# Logging et monitoring
python-json-logger==2.0.7

# Outils de développement
black==23.11.0
flake8==6.1.0

# Utilitaires supplémentaires
requests==2.31.0
python-dateutil==2.8.2
